requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-ckeditor==5.3.0
django-hosts==2.0
django-js-asset==0.1.1
django-modeltranslation==0.12.1
easy-thumbnails==2.4.1
idna==2.6
olefile==0.44
Pillow==4.2.1
python-akismet==0.3.0
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
